feat - Added new struct format for frames and threading for receiving frames
